requirements.txt
alembic==1.13.1
blinker==1.7.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.2
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
mysql-connector==2.2.9
mysql-connector-python==8.3.0
pycparser==2.21
PyMySQL==1.1.0
python-dotenv==1.0.1
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==3.0.1
